Move new networking stuff to a separate class
