+++ src/main/java/com/chailotl/inventory_sort/buttons/SortContainerButton.java
@@ -1,19 +1,18 @@
 package com.chailotl.inventory_sort.buttons;
 
-import com.chailotl.inventory_sort.Main;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.client.gui.screen.ingame.HandledScreen;
 import net.minecraft.client.gui.tooltip.Tooltip;
-import net.minecraft.network.PacketByteBuf;
 import net.minecraft.text.Text;
 
 /*? if >=1.20.5 {*/
-import net.minecraft.network.packet.CustomPayload;
-import net.minecraft.network.codec.PacketCodec;
+import com.chailotl.inventory_sort.Networking;
 /*?} else {*/
-/*import io.netty.buffer.Unpooled;
+/*import com.chailotl.inventory_sort.Main;
+import io.netty.buffer.Unpooled;
+import net.minecraft.network.PacketByteBuf;
 *//*?}*/
 
 @Environment(EnvType.CLIENT)
@@ -23,7 +22,7 @@ public class SortContainerButton extends InventoryButton
 	{
 		/*? if >=1.20.5 {*/
 		super(parent, x, y,  true, Icon.SORT, (button) ->
-				ClientPlayNetworking.send(new SortContainerPayload()));
+				ClientPlayNetworking.send(new Networking.SortContainerPayload()));
 		/*?} else {*/
 		/*super(parent, x, y,  true, Icon.SORT, (button) -> {
 			ClientPlayNetworking.send(Main.SORT_CONTAINER,
@@ -33,20 +32,4 @@ public class SortContainerButton extends InventoryButton
 
 		setTooltip(Tooltip.of(Text.translatable("gui.inventory_sort.tooltip.sort_container")));
 	}
-
-	/*? if >=1.20.5 {*/
-	public record SortContainerPayload() implements CustomPayload {
-		public static final CustomPayload.Id<SortContainerPayload> ID = new CustomPayload.Id<>(Main.SORT_CONTAINER);
-		public static PacketCodec<PacketByteBuf, SortContainerPayload> CODEC =
-				PacketCodec.ofStatic(SortContainerPayload::encode, SortContainerPayload::decode);
-
-		@Override
-		public Id<? extends CustomPayload> getId() {
-			return ID;
-		}
-
-		private static void encode(PacketByteBuf buf, SortContainerPayload self) {}
-		private static SortContainerPayload decode(PacketByteBuf buf) {return new SortContainerPayload();}
-	}
-	/*?}*/
 }
